7.e Add create new hero feature to HeroList and HeroDetail Components
add an Add button to HeroList html and bind to addHero() method

add addingHero property to HeroList and initialize as false and modify
onSelect(hero) method to set addingHero to false

add addHero() method to HeroList to set addingHero to true and
selectedHero to null

add close(hero) method to HeroList that sets addingHero to false and
calls getHeroes() method which re-calls HeroService

add <div *ngIf="addingHero"> to HeroList html that displays child
HeroDetailComponent and binds its close event to HeroList close(hero)
method.

add navigated property to HeroDetailComponent and initialize to false

modify ngOnInit() method in HeroDetail to now check for undefined route
parameter and set navigated to false and hero to a new Hero().

modify save() method in HeroDetail to set hero property to the returned
hero from the HeroService save() callback so that a new hero will have
correct id.

modify goBack(hero) method in HeroDetail to only use
window.history.back() if it was naviagted to,