7.h Add search feature to HeroDashboardComponent
create an rxjs-extensions.ts file with all needed RxJS imports and
import file into app.module.

create HeroSearchService that imports Hero, Http and Response from
@angular/core and Observable from RxJS and add a search(string) method
that builds a url to webApi with a query string containing the string
argument and returns an Observable<Hero>.

create HeroSearchComponent import and inject HeroSearchService and
Router, import Hero and also Observable and Subject from RxJS

add to HeroSearchComponent a heroes property as Observable<Hero[]> and
serachTerms property as Subject<string>. Subject produces an observable
event stream and searchTerms is an Observable<string>.

add to HeroSearchComponent a search(string) method that puts string
argument into searchTerms observable event stream by calling
next(string).

in HeroSearchComponent, implement ngOnInit() method that assigns to the
heroes property the return value of calls to HeroSearchService when
searchTerms has a value. The calls to HeroSearchSevice are throttled to
every 0.3 sec and only if there is a change in the searchTerms event
stream.

add to HeroSearchComponent a goToDetail(hero) method that routes to the
HeroDetailComponent of the hero argument.

add to HeroSearchComponent an html template that contains an input text
box with its keyup event bound to search(string) and an <div *ngFor >
that lists the heroes property after it flows through async pipe which
converts Observable<Hero[]> into Hero[]. Bind the <div> to hero.name and
bind the click event of the <div> to the goToDetail(hero) method.

add <hero-search> directive to HeroDashboardComponent